nixos/qemu-vm: fix diskless VMs
Previously, it was possible to run with a tmpfs / with
`virtualisation.diskImage = null;`, this was likely broken by my changes
in 4b4e4c3ef930c6106918d27a933de1b6965eff2c.

It is reintroduced by disabling properly the bootloader for now, as it
is complicated to make it work with.